Start with typing the ExperimentDict
