+++ src/sqooler/utils.py
@@ -7,7 +7,7 @@ import time
 import traceback
 import regex as re
 
-from .schemes import Spooler, ResultDict, StatusMsgDict
+from .schemes import Spooler, ResultDict, StatusMsgDict, ExperimentDict
 from .storage_providers import StorageProvider
 
 
@@ -117,3 +117,27 @@ def main(
         )
 
         counter += 1
+
+
+def create_memory_data(
+    shots_array: list, exp_name: str, n_shots: int
+) -> ExperimentDict:
+    """
+    The function to create memory key in results dictionary
+    with proprer formatting.
+    """
+    exp_sub_dict: dict = {
+        "header": {"name": "experiment_0", "extra metadata": "text"},
+        "shots": 3,
+        "success": True,
+        "data": {"memory": None},
+    }
+
+    exp_sub_dict["header"]["name"] = exp_name
+    exp_sub_dict["shots"] = n_shots
+    memory_list = [
+        str(shot).replace("[", "").replace("]", "").replace(",", "")
+        for shot in shots_array
+    ]
+    exp_sub_dict["data"]["memory"] = memory_list
+    return ExperimentDict(**exp_sub_dict)
